refactor: extract redux logics to hooks
